halt-workspace/requirements.txt
databricks-sdk==0.23.0
pandas==2.2.1
python-dotenv==1.0.1
requests==2.31.0
